--- app/blog/page.tsx
@@ -1,101 +1,0 @@
-"use client";
-
-import { useState, useEffect } from "react";
-import Link from "next/link";
-import { motion } from "motion/react";
-import { ArrowLeft } from "lucide-react";
-import { BlogCard } from "../../components/BlogCard";
-import { blogPosts } from "../../src/data/blog-posts";
-
-const categories = [
-  "All",
-  "Interview Prep",
-  "Competitive Programming",
-  "System Design",
-  "Success Stories",
-  "Career Tips",
-];
-
-export default function Blog() {
-  const [selectedCategory, setSelectedCategory] = useState("All");
-  const [filteredPosts, setFilteredPosts] = useState(blogPosts);
-
-  useEffect(() => {
-    if (selectedCategory === "All") {
-      setFilteredPosts(blogPosts);
-    } else {
-      setFilteredPosts(blogPosts.filter(post => post.category === selectedCategory));
-    }
-  }, [selectedCategory]);
-
-  const formatDate = (dateString: string) => {
-    const date = new Date(dateString);
-    return date.toLocaleDateString('en-US', { 
-      year: 'numeric', 
-      month: 'long', 
-      day: 'numeric' 
-    });
-  };
-
-  return (
-    <div className="bg-white min-h-screen">
-      {/* Back to Home Button */}
-      <motion.div className="max-w-7xl mx-auto px-6 pt-8 group">
-        <Link href="/" className="inline-flex items-center gap-1 text-indigo-600 hover:text-indigo-800 font-semibold text-sm mb-8 transition-all transform">
-          <ArrowLeft className="h-4 w-4 transition-transform duration-300 group-hover:-translate-x-1" />
-          Back to Home
-        </Link>
-      </motion.div>
-      {/* Header */}
-      <div className="bg-linear-to-r from-indigo-600 to-purple-600 text-white py-16">
-        <div className="mx-auto max-w-7xl px-6 lg:px-8">
-          <div className="mx-auto max-w-2xl text-center">
-            <h1 className="text-4xl font-bold tracking-tight sm:text-6xl">
-              CSL Blog
-            </h1>
-            <p className="mt-6 text-lg leading-8 text-indigo-100">
-              Insights, tips, and success stories from our community
-            </p>
-          </div>
-        </div>
-      </div>
-
-      {/* Category Filter */}
-      <div className="border-b border-gray-200">
-        <div className="mx-auto max-w-7xl px-6 lg:px-8 py-8">
-          <div className="flex flex-wrap gap-4 justify-center">
-            {categories.map((category) => (
-              <button
-                key={category}
-                onClick={() => setSelectedCategory(category)}
-                className={`px-4 py-2 rounded-full text-sm font-medium transition-colors ${
-                  selectedCategory === category
-                    ? 'bg-indigo-600 text-white'
-                    : 'bg-gray-100 text-gray-700 hover:bg-gray-200'
-                }`}
-              >
-                {category}
-              </button>
-            ))}
-          </div>
-        </div>
-      </div>
-
-      {/* Blog Posts */}
-      <div className="mx-auto max-w-7xl px-6 lg:px-8 py-16">
-        <div className="grid grid-cols-1 gap-8 lg:grid-cols-2 xl:grid-cols-3">
-          {filteredPosts.map((post) => (
-            <BlogCard key={post.id} post={post} className="w-full max-w-xl" href={`/blog/${post.slug}`} />
-          ))}
-        </div>
-
-        {filteredPosts.length === 0 && (
-          <div className="text-center py-16">
-            <h3 className="text-lg font-medium text-gray-900 mb-2">No posts found</h3>
-            <p className="text-gray-600">Try selecting a different category.</p>
-          </div>
-        )}
-      </div>
-    </div>
-  );
-}
