+++ src/app/page.tsx
@@ -1,12 +1,12 @@
-import Navbar from "@/components/Navbar";
-import CTA from "@/components/CTA";
+import Navbar from "@/components/layouts/navbar";
+import CTA from "@/components/sections/cta";
 import Hero from "@/components/sections/hero";
-import JoinUs from "@/components/JoinUs";
-import Services from "@/components/Services";
-import Team from "@/components/Team";
-import Discord from "@/components/Discord";
+import JoinUs from "@/components/join-us";
+import Services from "@/components/sections/services";
+import Team from "@/components/sections/team-info";
+import Discord from "@/components/discord";
 import Testimonials from "@/components/sections/testimonials";
-import Content from "@/components/Content";
+import Content from "@/components/contents";
 
 export default function HomePage() {
   return (
