+++ src/components/sections/team-info.tsx
